Refactor mechaclient class to mechaclient.py at project root. Fixed misc. imports to prevent circular loading after refactor.
